NO-JIRA: Deposit filter for last modified timestamp (#31)
* create an SQL VIEW for the "last modified timestamp" of a deposit
* replace the existing lastModified SQL query with that VIEW
* replace the existing earlierThan, laterThan and atTimestamp GraphQL filters (which actually were about the "creation timestamp") with createdEarlierThan, createdLaterThan and createdAtTimestamp GraphQL filters
* add additional GraphQL filters for lastModfiedEarlierThan, lastModfiedLaterThan and lastModfiedAtTimestamp
* add the "last modified filters" to the SQL query generator

In easy-manage-deposit the reports over the last 24 hours are currently generated by looking at the last modified timestamp of the deposit.properties files. Only those deposits that were either added or changed in the last 24 hours get added to the report.
In replacing the deposit.properties files with the easy-deposit-properties database/service, we want to preserve this functionality and require "last modified filters".